カルーセルウィジェットの分割
